+++ CHANGELOG.md
@@ -21,7 +21,8 @@ Stages are now the central part; multiple project can be associated with the sam
 
 * stool config changes
   * dumped `diskMin`
-  
+  * dumped `macros`
+q  
 * stage config changes
   * dumped `autoRefresh`